Code to Run
The code sets up a connection to the display modules and then enters an endless loop. Inside this loop, it uses the show_message() function from the luma library to send the text "Hello" to the display, which then scrolls across the screen. You can easily change this message to whatever you'd like